fix(KvWwwHeaderBasic): align desktop bar to Figma
- Move logo into LinkBar: in-flow far-left at md+, centered on mobile
- Drop ring/border styling from Lend to match About
- Remove 'Partner with us' from About menu (now a top-level nav link)
